pyproject.toml
[project]
name = "algorithm-visualizations"
version = "0.1.0"
description = "A visualizer of some algorithms"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.6",
]

[project.scripts]
visualizations = "visualizations.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true
